setup/repo/forensic-vm-1.0/forensicVM/main/requirements.txt
altgraph==0.17.3
asgiref==3.6.0
attrs==22.2.0
auto-py-to-exe==2.30.0
Automat==22.10.0
bottle==0.12.23
bottle-websocket==0.2.9
certifi==2022.12.7
cffi==1.15.1
chardet==4.0.0
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
cryptography==39.0.1
Django==4.1.9
django-extensions==3.2.1
django-revproxy @ https://github.com/jazzband/django-revproxy/archive/refs/tags/0.11.0.zip#sha256=182206aa1d1719aa5abf7379efb45ccb93c0d7088b8a25d8a69c67459c1d547e
django-session-timeout==0.1.0
django-upgrade==1.13.0
djangorestframework==3.14.0
docopt==0.6.2
Eel==0.14.0
future==0.18.3
gevent==22.10.2
gevent-websocket==0.10.1
gpg===1.14.0-unknown
greenlet==2.0.2
guibot==0.50.1
gunicorn==20.1.0
httplib2==0.18.1
hyperlink==21.0.0
idna==3.4
imagemounter==3.1.0
incremental==22.10.0
iotop==0.6
IProgress==0.4
MarkupSafe==2.1.2
MouseInfo==0.1.3
numpy==1.24.2
pefile==2023.2.7
Pillow==9.5.0
pipreqs==0.4.11
psutil==5.9.5
PyAutoGUI==0.9.53
pycparser==2.21
pycurl==7.43.0.6
pygame==2.1.3
PyGetWindow==0.0.9
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyMsgBox==1.0.9
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySide2==5.15.2.1
PySimpleGUI==4.60.4
PySimpleGUIDesigner==0.1.4.7
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
python3-xlib==0.15
pytweening==1.0.4
pytz==2023.3
PyWebDAV3==0.10.0
pywin32-ctypes==0.2.0
PyYAML==5.3.1
qemu==0.0.0a1
qemu.qmp==0.0.2
qmp==0.0.1
reportbug==7.10.3+deb11u1
requests==2.31.0
shiboken2==5.15.2.1
six==1.16.0
sqlparse==0.4.4
tabulate==0.9.0
termcolor==2.2.0
tokenize-rt==5.0.0
tqdm==4.64.1
Twisted==22.10.0
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
vnc2flv-rec==0.2.0
vncdotool==1.0.0
websockify==0.10.0
Werkzeug==2.2.3
whichcraft==0.6.1
yarg==0.1.9
zope.event==4.6
zope.interface==5.5.2
